At this stage of the program, we create data using global new and delete (alongisde class new/elete) With the node system, upon creting a node, we set the address and the block of memory.
The list of our nodes does not correlate with the addresses ofthe nodes, what do I mean?

We create node_1 first, we then create node_2 however this is placed at the front, making node_2 the new head.

If we look at the memory address, while both are stored contigously with padding inbetween, node_2 is infront of node_1.

While we can walk through the heap, the changing of the sequence of nodes doesnt correlate to the addresses also changing.

If we was to but a node inbetween node1/2, the heap will store contigously in this order:

node_1 -> node_2 - > node_3

However the node sequence is this.

node_2 -> node_3 -> node_1

I dont think this will be an issue fully, with the current naming convention it can certainly an issue for readibility as the names imply a numerical sequence.

        * current selfawareness mitigates this, however if I forget, theres problems.

Nevertheless we are walking through the heap now, maybe this slight issue will change once I incooperate into project?

Im already thinking of ideas however cannot write them at this moment.